+++ torcf/__init__.py
@@ -13,7 +13,6 @@ from .creds import get_credentials
 
 # Globals
 from .globals import Globals, GlobalVars
-from .globals import ToRPost
 
 # Logger
 from .logger import Log
@@ -22,10 +21,7 @@ from .logger import Log
 from .main import clone_finder
 
 # Post manipulation
-from .post import add_post
-from .post import check_post
-from .post import check_mod_log
+from .post import ClonedPost
+from .post import PostHandler
+from .post import ToRPost
 from .post import check_mod_queue
-from .post import duplicate
-from .post import find_wanted
-from .post import update_post_list