+++ torcf/globals.py
@@ -16,35 +16,55 @@
 
     Contact me at murdomaclachlan@duck.com
 """
-from praw.models import Submission
+from os import environ, makedirs
+from os.path import expanduser, isdir
 from smooth_logger import Logger
-from typing import Iterable, List, Union, Any
+from sys import platform
+from typing import List, Dict
 
 
-VERSION = "1.0.0-dev41-20221005"
+VERSION = "1.0.0-dev42-20221005"
 
 
 class GlobalVars:
     """Container and handler for global variables used throughout the program.
     """
-    def __init__(self: object):
-        # Attributes declared here as None will be properly initialised later in the
-        # run-time, depending on what arguments are passed
-        self.CHECK_FOR_SUB = None
-        self.MODLOG = None
-        self.MODQUEUE = None
-        self.REMOVE = None
-        self.SUBREDDITS = None
-        self.VERBOSE = None
-        self.WAIT = None
-        self.removed_posts = None
-        self.wanted_posts = None
-        self.wanted_posts_last = None
-        # Attributes declared here should have constant initial values
-        self.first_post_url = ""
-        self.posts = []
+    def __init__(self):
+        self.check_for_sub = False
+        self.modlog = False
+        self.modqueue = False
+        self.remove: Dict[str, bool] = {}
+        self.subreddits = False
+        self.verbose = False
+        self.wait = 30
+        self.__get_config_path()
+        self.Handler = None
 
-    def __determine_wait(self, argv: List, Log: Logger) -> None:
+    def __conf_remove(self, argv: List[str], Log: Logger) -> None:
+        """
+        """
+        if "--remove" in argv or "-r" in argv:
+            index = (
+                argv.index("--remove")
+                if "--remove" in argv else
+                argv.index("-r")
+            )
+            try:
+                kind = argv[index+1]
+                self.remove["flaired"] = True if kind in ["all", "flaired"] else False
+                self.remove["unflaired"] = (
+                    True if kind in ["all", "unflaired"] else False
+                )
+            except IndexError:
+                Log.new(
+                    f"{argv[index]} was passed with no subsequent kind; defaulting to"
+                    + " all.",
+                    "WARNING"
+                )
+                self.remove["flaired"] = True
+                self.remove["unflaired"] = True
+
+    def __conf_wait(self, argv: List, Log: Logger) -> None:
         """Determine the number of seconds TCF should wait between cycles. Default to
         30 if no valid value is passed.
 
@@ -61,58 +81,47 @@ class GlobalVars:
                 argv.index("-w")
             )
             try:
-                self.WAIT = abs(int(argv[index+1])); return
+                self.wait = abs(int(argv[index+1])); return
             except (IndexError, ValueError):
                 Log.new(
                     f"{argv[index]} was passed with no subsequent time value;"
                     + " defaulting to 30.",
                     "WARNING"
                 )
-        self.WAIT = 30
+        self.wait = 30
 
-    def check_skip(self, post_list: Iterable) -> bool:
-        """Using the first_post_url value, check whether we should skip this
-        cycle.
+    def __get_config_path(self) -> None:
+        """Detects OS and defines the appropriate save path for config and logs. Exits on
+        detecting an unsupported OS. Supported OSes are: Linux, MacOS, Windows.
 
-        :param post_list: the list of posts in the queue
-
-        :return: a boolean success status
+        :return: A string dictionary containing the newly defined save paths.
         """
-        for post in post_list:
-            post_url = f"https://reddit.com{post.permalink}"
-            # If the first post in the queue is the same as it was last time we checked,
-            # nothing has changed and we skip this cycle
-            if self.first_post_url == post_url:
-                return True
-            # Else, update the URL of the first post and continue to the main checks
-            else:
-                self.first_post_url = post_url
-                return False
+        home = expanduser("~")
+        os = "".join(list(platform)[:3])
 
-    def clean(self) -> None:
-        """Deletes all currently stored members of self.posts.
-        """
-        del self.posts[:]
-        del self.removed_posts[:]
+        # Route for a supported operating system
+        if os in ["dar", "lin", "win"]:
+            path = (
+                environ[
+                    "APPDATA"] + "\\torcf" if os == "win" else f"{home}/.config/torcf"
+            )
+            # Create any missing directories
+            if not isdir(path):
+                print(f"Making path: {path}")
+                makedirs(path, exist_ok=True)
+            self.config_path = path
 
-    def delete_post(self, permalink):
-        """Find and delete a single member of self.posts using a given permalink.
-
-        :param permalink: The permalink of the post to delete.
-        """
-        for i in self.posts:
-            if i.permalink == permalink:
-                self.posts.remove(i)
+        # Exit if the operating system is unsupported
+        else:
+            print(f"FATAL: Unsupported operating system: {os}, exiting.")
+            exit()
 
     def get_subs(self) -> None:
         """Get, from user input, a list of subreddits to search for posts from.
         """
-        self.SUBREDDITS = input(
+        self.subreddits = input(
             "Please enter the subreddits to search for, separated by spaces.\n  >> "
         ).casefold().split(" ")
-        self.removed_posts = []
-        self.wanted_posts = []
-        self.wanted_posts_last = [ToRPost(None, dummy=True)]
 
     def process_args(self, argv: List[str], Log: Logger) -> None:
         """Process any passed runtime arguments.
@@ -120,63 +129,11 @@ class GlobalVars:
         :param argv: the list of runtime arguments
         :param Log: The Logger object.
         """
-        self.CHECK_FOR_SUB = ("--check" in argv or "-c" in argv)
-        self.MODLOG = ("--modlog" in argv or "-l" in argv)
-        self.MODQUEUE = ("--modqueue" in argv or "-q" in argv)
-        self.REMOVE = ("--remove" in argv or "-r" in argv)
-        self.VERBOSE = ("--verbose" in argv or "-v" in argv)
-        self.__determine_wait(argv, Log)
-
-    def wanted_posts_changed(self) -> bool:
-        """Check if there has been any change to the tracked posts from monitored
-        subreddits.
-
-        :return: True if there has been a change; else False.
-        """
-        if len(self.wanted_posts) > 0:
-            for i in range(len(self.wanted_posts)):
-                if self.wanted_posts[i].flair != self.wanted_posts_last[i].flair:
-                    self.wanted_posts_last = self.wanted_posts
-                    return True
-        return False
-
-
-class ToRPost:
-    """Represents an instance of a post on the r/TranscribersOfReddit subreddit, with
-    several attributes and methods which help the program investigate and manipulate the
-    post.
-    """
-    def __eq__(self, other: Union[Submission, Any]):
-        return self.orig_link == other.orig_link
-
-    def __init__(self, praw_obj: Union[Submission, Any], dummy: bool = False):
-        self.praw_obj = None
-        self.created = None
-        self.flair = ""
-        self.orig_link = None
-        self.permalink = None
-        self.subreddit = None
-        if not dummy:
-            self.praw_obj = praw_obj
-            self.created = self.praw_obj.created_utc
-            self.flair = self.praw_obj.link_flair_text
-            self.orig_link = self.praw_obj.url
-            self.permalink = self.praw_obj.permalink
-            self.subreddit = self.praw_obj.title.split(" |")[0].casefold()
-
-    def update_flair(self, new_flair: str) -> None:
-        """Updates the currently stored flair of the ToRPost with a new, given one.
-
-        :param new_flair: string
-        """
-        self.flair = new_flair
-
-    def remove(self, GlobalHandler: GlobalVars) -> None:
-        """Removes the post from the ToR queue. Can only be invoked by moderators of
-        r/TranscribersOfReddit; will crash the program if attempted by anyone else.
-        """
-        self.praw_obj.mod.remove(mod_note="Cloned post. (automatic removal)")
-        GlobalHandler.delete_post(self.permalink)
-
+        self.check_for_sub = ("--check" in argv or "-c" in argv)
+        self.modlog = ("--modlog" in argv or "-l" in argv)
+        self.modqueue = ("--modqueue" in argv or "-q" in argv)
+        self.verbose = ("--verbose" in argv or "-v" in argv)
+        self.__conf_wait(argv, Log)
+        self.__conf_remove(argv, Log)
 
 Globals = GlobalVars()